// src/types.ts
export interface IPaginationInfo {
    count: number;
    pages: number;
    next: string | null;
    prev: string | null
};
  
export interface IRickAndMorty {
    id: string;
    name: string;
    status: string;
    species: string;
    type: string;
    gender: string;
    origin: {
        name: string;
        url: string;
    };
    location: {
        name: string;
        url: string;
    };
    image: string;
    episode: string[];
    url: string;
    created: string
};

export interface IResponse {
    info: IPaginationInfo,
    results: IRickAndMorty[]
};